+++ priv/migrations/005_create-character_reports.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS character_reports (
   id INTEGER PRIMARY KEY AUTOINCREMENT,
   session_id INTEGER,
   character_id INTEGER,
-  xp_gained REAL,
+  xp_gained INTEGER,
   total_xp REAL,
   level_up BOOLEAN,
   created_at TEXT,
